+++ src/test/java/racingcar/RacingCarTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.ui.ResultView;
 
 import java.util.Arrays;
 import java.util.List;
